refactor: move collision to a separate file
