+++ frontend/src/game/levels/types/index.ts
@@ -1,6 +1,8 @@
 import LevelId from './levelIds';
 import LevelLoad from './levelLoad';
+import Position from './position';
+import Surface from './surface';
 
 export {
-  LevelId, LevelLoad,
+  LevelId, LevelLoad, Surface, Position,
 };
