+++ frontend/src/game/levels/types/position.ts
@@ -1,0 +1,3 @@
+type Position = { position:Line };
+
+export default Position;
