+++ frontend/src/game/levels/types/surface.ts
@@ -1,0 +1,6 @@
+import { Line } from '../../shapes';
+import SurfaceType from '../../types/surfaceType';
+
+type Surface = { type:SurfaceType, platform:boolean, position:Line, angle:number };
+
+export default Surface;
